WeaponElongator proof of concept.
All Melee Weapons now have a part that checks for Elongated Paws when the wepon is equipped, and if it finds them, adds the DamageBonus (StrMod/2) to the MaxStrCap.

The part will also perform this check on Unequip and remove the bonus.

Couldn't get it to work on StatChangeEvent via the part, but the part can be called from the mutation itself during a StatChangeEvent and that seems to be doing it.