FInished Refactoring.
Mutations can now either inherit from BaseManagedDefaultEquipmentMutation and utilise overrides for getting the properties of the included NaturalWeapon class or they can be "wrapped" with the IDefaultNaturalWeaponManager interface and implement the required methods.

The above, used in conjunction with a Mod that inherits from ModNaturalWeaponBase<T>, where T is the Mutation in question, can have its NaturalWeapon class properties added on top of the default fist during the mutation's OnRegenerateDefaultEquipment call by applying the Modification to the otherwise unchanged default fist.

This refactoing also included altering the HarmonyPatches.cs file to trial "wrapping" Burrowing Claws and Crystallinity. It works, but Crystallinity needed some attention since it's added explicitly by the meal, instead of by reference to its mutation entry.

Added several new HelperMethods, including a targeted .Trace() Debug method for MeleeWeapon parts which will serve as a template for future ones.

Prefixed the Mod's namespace and made some adjustments to it to better distinguish between it and the eponymous mutation.