Added constructor for INaturalWeapon Part which takes an INaturalWeapon and assigns the new one each of its values, creating a distinct but identical copy.
Adjusted the Gigantified ObjectBuilder to make it utilise its Part counterpart in the process of processing its wish (created objects are now considered Gigantic on their way out of the factory, instead of after leave it).

Adjusted InventoryGigantifier to stick around if it's attached to a merchant. It has some rules for gigantifying merchant inventories when stocking/restocking.
- yet to be tested.

Some work on fixing the bug wherein Fugue clones equipped with a GiganticExoframe aren't conferred Gigantism. Currently still broken.
- created a weapon which should assist in testing.