added harmony patch to ModGigantic to hopefully make gigantic objects that become animated to go through gigantification in the process.
added missed serialization to GigantismPlus' "Close Fist" activated ability.

minor adjustment ElongatedPaws' comments.

added some static/instance methods to Gigantified, as well as some contextual processing for true kin.

created a "fixed" version of horns that is, for all intents and purposes, the same as vanilla, but inherits `BaseDefaultEquipmentMutation`, in the hopes that that somewhat assists in it playing nice with the NaturalEquipmentManager part.