Reworked how NaturalEquipmentMods are collected by the NaturalWeaponManager.
It's not quite as straight forward as before, but it's doing a lot more under the hood now to lessen the necessity for serializing.

Adjustments aren't applying correctly, but the weapon damage chanes are. Descriptions are working well, too.

Need to add tracing to Adjustment.Apply() soI can see why it's failing.