Changed the ManageDefaultEquipmentEvent set to inherit from a common base class. Tidied it way up in the process.
Muted the shaders a bit so that they're a little less toward the orange side of fleshy after feedback from gnarf (a CoQ game dev).

Freshened up the WeaponElongater which should hopefully now display the adjustments that are being made to it (to be tested).

Tightened up the sending and handling of the DescribeModificationEvents where NaturalEquipmentMods are concerned.