Wrassle-gear now successfully takes on the melee weapon properties of any default equipment that occupies its intended slot.
Reduced some of the debug spam that's showing up with a lot of it turned off now.

Added context argument to some of the SeededRandom extension methods to braoden the variety without compromising the "certainty".

Added checks so that you're alerted about trying to vault over things that either flying or yourself.

Adjusted most of the vaulting parts to check for PlayerControlled instead of Player to get more consistent behaviour.