added a field to IConditions to determin how to treat a null subject. Default behaviour is to return true.
Added some static pre-configured IConditions to a static class to serve as an example and to reduce how much code is being doubled up on.

Adjustes seceral Creatures' weight to be 1/5th of it's xml entry so that when it's made gigantic it doesn't weigh as much as a small island.