Added a new set of events for checking the giganticness of an object, split into Creature and Equipment to mirror the base-game's nomenclature/ontology.
Changed GigantismPlus' mutation entry name to "Gigantism" so that it's the mutation that comes up "by entry name", and so that other mods wanting to replace Gigantism's entry won't run into the exact same issues these changes are designed to fix.

fixed some issues with some of the population tables.